// packages/core/src/schemas/dml-lockedCanvas.ts
/**
 * Auto-generated from dml-lockedCanvas.xsd
 * DO NOT EDIT - Generated by xsd-converter
 */

import type { XsdSchema, XsdSimpleType, XsdComplexType, XsdElement, XsdGroup, XsdAttributeGroup } from '../types'

export const dmlLockedCanvasSchema: XsdSchema = {
  "targetNamespace": "http://purl.oclc.org/ooxml/drawingml/lockedCanvas",
  "imports": [
    {
      "namespace": "http://purl.oclc.org/ooxml/drawingml/main",
      "schemaLocation": "dml-main.xsd"
    }
  ],
  "simpleTypes": [],
  "complexTypes": [],
  "elements": [
    {
      "name": "lockedCanvas",
      "type": {
        "kind": "external",
        "prefix": "a",
        "name": "CT_GvmlGroupShape"
      },
      "minOccurs": 1,
      "maxOccurs": 1,
      "nillable": false
    }
  ],
  "groups": [],
  "attributeGroups": []
}

export default dmlLockedCanvasSchema